// src/app/Shared/Interfaces/Cart/icart.ts
 export interface ICart {
  _id: string;
  cartOwner: string;
  products: Products[];
  createdAt: string;
  updatedAt: string;
  __v: number;
  totalCartPrice: number;
}

 export interface Products {
  count: number;
  _id: string;
  product: Product;
  price: number;
}

 export interface Product {
  subcategory: Subcategory[];
  _id: string;
  title: string;
  quantity: number;
  imageCover: string;
  category: Category;
  brand: Category;
  ratingsAverage: number;
  id: string;
}

 export interface Category {
  _id: string;
  name: string;
  slug: string;
  image: string;
}

 export interface Subcategory {
  _id: string;
  name: string;
  slug: string;
  category: string;
}